fix(tests/shell): fix build
