+++ src/shell/CMakeLists.txt
@@ -2,8 +2,10 @@ add_executable(lean lean.cpp json.cpp server.cpp)
 target_link_libraries(lean leanstatic)
 install(TARGETS lean DESTINATION bin)
 
-add_executable(lean_js lean_js.cpp)
+add_executable(lean_js lean_js.cpp lean_js_main.cpp)
 target_link_libraries(lean_js leanstatic)
+
+add_library(shell_js OBJECT lean_js.cpp)
 
 if(${EMSCRIPTEN})
     target_link_libraries(lean "--memory-init-file 0")
