+++ src/shell/lean_js.cpp
@@ -81,23 +81,3 @@ int emscripten_import_module(std::string mname) {
 int emscripten_process_file(std::string input_filename) {
     return g_shell->process_file(input_filename);
 }
-
-#if defined(LEAN_EMSCRIPTEN)
-
-#include <emscripten/bind.h>
-EMSCRIPTEN_BINDINGS(LEAN_JS) {
-        emscripten::function("lean_init", &initialize_emscripten);
-        emscripten::function("lean_import_module", &emscripten_import_module);
-        emscripten::function("lean_process_file", &emscripten_process_file);
-}
-
-int main() { return 0; }
-
-#else
-
-int main(int, char **) {
-    initialize_emscripten();
-    emscripten_import_module("standard");
-}
-
-#endif
